+++ admin/tabs/server/websocket/WebSocketModel.ts
@@ -4,6 +4,7 @@
  *
  * Copyright © 2023 Extremely Heavy Industries Inc.
  */
+import {exportFilenameWithDate} from '@xh/hoist/admin/AdminUtils';
 import * as Col from '@xh/hoist/admin/columns';
 import {GridModel} from '@xh/hoist/cmp/grid';
 import {div, p} from '@xh/hoist/cmp/layout';
@@ -12,7 +13,7 @@ import {textInput} from '@xh/hoist/desktop/cmp/input';
 import {Icon} from '@xh/hoist/icon';
 import {makeObservable, observable, runInAction} from '@xh/hoist/mobx';
 import {Timer} from '@xh/hoist/utils/async';
-import {LocalDate, SECONDS} from '@xh/hoist/utils/datetime';
+import {SECONDS} from '@xh/hoist/utils/datetime';
 import {isDisplayed} from '@xh/hoist/utils/js';
 import {isEmpty} from 'lodash';
 import {createRef} from 'react';
@@ -37,7 +38,7 @@ export class WebSocketModel extends HoistModel {
         this.gridModel = new GridModel({
             emptyText: 'No clients connected.',
             enableExport: true,
-            exportOptions: {filename: `${XH.appCode}-ws-connections-${LocalDate.today()}`},
+            exportOptions: {filename: exportFilenameWithDate('ws-connections')},
             selModel: 'multiple',
             store: {
                 idSpec: 'key',